FEAT: Packaging timing benchmarking + v2.5.1 improvements :speedboat:
